created log out function and user authorities to edit and delete article

+++ classes/Ijdb/controllers/Joke.php
@@ -2,16 +2,18 @@
 
 namespace Ijdb\Controllers;
 use \FrameWork\DatabaseTable;
+use \FrameWork\Authentication;
 
 
 class Joke{
     private $authorsTable;
     private $jokesTable;
 
-    public function __construct( DatabaseTable $jokesTable, DatabaseTable $authorsTable )
+    public function __construct( DatabaseTable $jokesTable, DatabaseTable $authorsTable, Authentication $authentication )
     {
         $this -> jokesTable = $jokesTable;
         $this -> authorsTable = $authorsTable;
+        $this -> authentication = $authentication;
     }
 
     public function list() {
@@ -27,7 +29,8 @@ class Joke{
                 'joketext' => $joke['joketext'],
                 'jokedate' => $joke['jokedate'],
                 'name' => $author['name'],
-                'email' => $author['email']
+                'email' => $author['email'],
+                'authorId' => $author['id']
             ];
 
         }
@@ -37,11 +40,14 @@ class Joke{
 
         $totalJokes = $this -> jokesTable -> total();
 
+        $author = $this -> authentication -> getUser();
+
             return ['template' => 'jokes.html.php', 
                      'title' => $title,
                      'variables' => [
                         'totalJokes' => $totalJokes,
-                        'jokes' => $jokes
+                        'jokes' => $jokes,
+                        'userId' => $author['id'] ?? null
                      ]
                     ];
     }
@@ -53,15 +59,31 @@ class Joke{
     }
 
     public function delete() {
+        $author = $this -> authentication -> getUser();
+
+        $joke = $this -> jokesTable -> findById($_GET['id']);
+        if( $joke['authorId'] != $author['id']){
+            return;
+        }
+
         $this -> jokesTable -> delete( $_POST{'id'});
 
         header('location: /joke/list');
     }
 
     public function saveEdit() {
+        $author = $this -> authentication -> getUser();
+
+        if(isset($_GET['id'])){
+            $joke = $this -> jokesTable -> findById($_GET['id']);
+            if( $joke['authorId'] != $author['id']){
+                return;
+            }
+        }
+
         $joke = $_POST['joke'];
         $jokep['jokedate'] = new \DateTime();
-        $joke['authorId'] = 1;
+        $joke['authorId'] = $author['id'];
 
         $this -> jokesTable -> save($joke);
 
@@ -69,6 +91,8 @@ class Joke{
     }
 
     public function edit() {
+        $author = $this -> authentication -> getUser();
+
         if(isset($_GET['id'])){
             $joke = $this -> jokesTable -> findById($_GET['id']);
         }
@@ -78,7 +102,8 @@ class Joke{
             return ['template' => 'editjoke.html.php', 
                     'title' => $title,
                     'variables' => [
-                        'joke' => $joke ?? null
+                        'joke' => $joke ?? null,
+                        'userId' => $author['id'] ?? null
                     ]
                     
                 ];
